+++ include/variables.h
@@ -259,3 +259,7 @@ struct unkEvent{  //figure out what this is later
 };
 extern unkEvent lbl_803A7288;
 extern u8 lbl_8037B0D0[0x8];
+
+class JKRThread;
+extern JSUList<JKRThread> lbl_8043428C;   // JSUList<JKRThread> JKRThread::sThreadList
+
